+++ wgnetstack/wgnetstack.go
@@ -6,7 +6,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	mathrand "math/rand/v2"
 	"net"
 	"net/netip"
 	"os"
@@ -20,6 +19,7 @@ import (
 	"github.com/fosrl/newt/logger"
 	"github.com/fosrl/newt/netstack2"
 	"github.com/fosrl/newt/proxy"
+	"github.com/fosrl/newt/util"
 	"github.com/fosrl/newt/websocket"
 	"golang.zx2c4.com/wireguard/device"
 	"golang.zx2c4.com/wireguard/tun"
@@ -116,49 +116,6 @@ func (s *WireGuardService) RemoveProxyTarget(proto, listenIP string, port int) e
 	return s.proxyManager.RemoveTarget(proto, listenIP, port)
 }
 
-// find an available UDP port in the range [minPort, maxPort] and also the next port for the wgtester
-func FindAvailableUDPPort(minPort, maxPort uint16) (uint16, error) {
-	if maxPort < minPort {
-		return 0, fmt.Errorf("invalid port range: min=%d, max=%d", minPort, maxPort)
-	}
-
-	// We need to check port+1 as well, so adjust the max port to avoid going out of range
-	adjustedMaxPort := maxPort - 1
-	if adjustedMaxPort < minPort {
-		return 0, fmt.Errorf("insufficient port range to find consecutive ports: min=%d, max=%d", minPort, maxPort)
-	}
-
-	// Create a slice of all ports in the range (excluding the last one)
-	portRange := make([]uint16, adjustedMaxPort-minPort+1)
-	for i := range portRange {
-		portRange[i] = minPort + uint16(i)
-	}
-
-	// Fisher-Yates shuffle to randomize the port order
-	for i := len(portRange) - 1; i > 0; i-- {
-		j := mathrand.IntN(i + 1)
-		portRange[i], portRange[j] = portRange[j], portRange[i]
-	}
-
-	// Try each port in the randomized order
-	for _, port := range portRange {
-		// Check if port is available
-		addr1 := &net.UDPAddr{
-			IP:   net.ParseIP("127.0.0.1"),
-			Port: int(port),
-		}
-		conn1, err1 := net.ListenUDP("udp", addr1)
-		if err1 != nil {
-			continue // Port is in use or there was an error, try next port
-		}
-
-		conn1.Close()
-		return port, nil
-	}
-
-	return 0, fmt.Errorf("no available consecutive UDP ports found in range %d-%d", minPort, maxPort)
-}
-
 func NewWireGuardService(interfaceName string, mtu int, generateAndSaveKeyTo string, host string, newtId string, wsClient *websocket.Client, dns string) (*WireGuardService, error) {
 	var key wgtypes.Key
 	var err error
@@ -190,7 +147,8 @@ func NewWireGuardService(interfaceName string, mtu int, generateAndSaveKeyTo str
 	}
 
 	// Find an available port
-	port, err := FindAvailableUDPPort(49152, 65535)
+	port, err := util.FindAvailableUDPPort(49152, 65535)
+
 	if err != nil {
 		return nil, fmt.Errorf("error finding available port: %v", err)
 	}
@@ -237,7 +195,7 @@ func NewWireGuardService(interfaceName string, mtu int, generateAndSaveKeyTo str
 
 	// Create the holepunch manager with ResolveDomain function
 	// We'll need to pass a domain resolver function
-	service.holePunchManager = holepunch.NewManager(sharedBind, newtId)
+	service.holePunchManager = holepunch.NewManager(sharedBind, newtId, "newt")
 
 	// Register websocket handlers
 	wsClient.RegisterHandler("newt/wg/receive-config", service.handleConfig)
